@@ -27,23 +27,24 @@
     "@angular/common": "^15.2.0",
     "@angular/compiler": "^15.2.0",
     "@angular/core": "^15.2.0",
-    "@angular/forms": "^15.2.0",
-    "@angular/localize": "^15.2.0",
-    "@angular/material": "^15.2.0",
+    "@angular/forms": "^15.2.7",
+    "@angular/localize": "^15.2.7",
+    "@angular/material": "^15.2.7",
     "@angular/platform-browser": "^15.2.0",
-    "@angular/platform-browser-dynamic": "^15.2.0",
-    "@angular/router": "^15.2.0",
-    "@supabase/supabase-js": "^2.7.1",
+    "@angular/platform-browser-dynamic": "^15.2.7",
+    "@angular/router": "^15.2.7",
+    "@dlr-shepard/shepard-client": "2023.4.11",
+    "@supabase/supabase-js": "^2.20.0",
     "node-sass-json-importer": "^4.3.0",
     "rxjs": "~7.8.0",
-    "sass": "^1.58.0",
+    "sass": "^1.62.0",
     "snakecase-keys": "^5.4.5",
     "tslib": "^2.5.0",
     "yaml": "^2.2.1",
     "zone.js": "^0.12.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^15.2.0",
+    "@angular-devkit/build-angular": "^15.2.6",
     "@angular-eslint/builder": "15.2.2-alpha.0",
     "@angular-eslint/eslint-plugin": "15.2.2-alpha.0",
     "@angular-eslint/eslint-plugin-template": "15.2.2-alpha.0",
@@ -51,16 +52,16 @@
     "@angular-eslint/template-parser": "15.2.2-alpha.0",
     "@angular/cli": "~15.2.0",
     "@angular/compiler-cli": "^15.2.0",
-    "@openapitools/openapi-generator-cli": "^2.5.2",
+    "@openapitools/openapi-generator-cli": "^2.6.0",
     "@types/jasmine": "~4.3.1",
     "@types/snakecase-keys": "^3.1.0",
-    "@typescript-eslint/eslint-plugin": "^5.51.0",
-    "@typescript-eslint/parser": "^5.51.0",
-    "cypress": "^12.5.1",
-    "eslint": "^8.34.0",
-    "eslint-config-prettier": "^8.6.0",
+    "@typescript-eslint/eslint-plugin": "^5.58.0",
+    "@typescript-eslint/parser": "^5.58.0",
+    "cypress": "^12.9.0",
+    "eslint": "^8.38.0",
+    "eslint-config-prettier": "^8.8.0",
     "eslint-plugin-import": "^2.27.5",
-    "eslint-plugin-jsdoc": "^40.0.0",
+    "eslint-plugin-jsdoc": "^40.3.0",
     "eslint-plugin-prefer-arrow": "^1.2.3",
     "eslint-plugin-rxjs-angular": "^2.0.0",
     "husky": "^8.0.3",
@@ -70,7 +71,7 @@
     "karma-coverage": "~2.2.0",
     "karma-jasmine": "~5.1.0",
     "karma-jasmine-html-reporter": "~2.0.0",
-    "prettier": "^2.8.4",
+    "prettier": "^2.8.7",
     "typescript": "~4.9.5"
   }
 }
